Extract some types into files
