+++ src/Elastic.Clients.Elasticsearch/Client/ElasticsearchClient-Manual.cs
@@ -17,6 +17,12 @@ namespace Elastic.Clients.Elasticsearch
 			return DoRequest<IndexRequestDescriptor<TDocument>, IndexResponse>(descriptor);
 		}
 
+		public Task<IndexResponse> IndexAsync<TDocument>(TDocument document, CancellationToken cancellationToken = default)
+		{
+			var descriptor = new IndexRequestDescriptor<TDocument>(documentWithId: document);
+			return DoRequestAsync<IndexRequestDescriptor<TDocument>, IndexResponse>(descriptor);
+		}
+
 		public Task<IndexResponse> IndexAsync<TDocument>(TDocument document, Action<IndexRequestDescriptor<TDocument>> configureRequest, CancellationToken cancellationToken = default)
 		{
 			var descriptor = new IndexRequestDescriptor<TDocument>(documentWithId: document);
