+++ src/Elastic.Clients.Elasticsearch/Common/Infer/Fields/Fields.cs
@@ -15,96 +15,6 @@ using Elastic.Transport;
 
 namespace Elastic.Clients.Elasticsearch;
 
-internal sealed class FieldsConverter : JsonConverter<Fields>
-{
-	public override Fields? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-	{
-		if (reader.TokenType == JsonTokenType.String)
-		{
-			Fields fields = reader.GetString();
-			return fields;
-		}
-		else if (reader.TokenType == JsonTokenType.StartArray)
-		{
-			var fields = new List<Field>();
-			while (reader.Read() && reader.TokenType != JsonTokenType.EndArray)
-			{
-				var field = JsonSerializer.Deserialize<Field>(ref reader, options);
-				fields.Add(field);
-			}
-			return new Fields(fields);
-		}
-
-		reader.Read();
-		return null;
-	}
-
-	public override void Write(Utf8JsonWriter writer, Fields value, JsonSerializerOptions options)
-	{
-		if (value is null)
-		{
-			writer.WriteNullValue();
-			return;
-		}
-
-		writer.WriteStartArray();
-		foreach (var field in value.ListOfFields)
-		{
-			JsonSerializer.Serialize(writer, field, options);
-		}
-		writer.WriteEndArray();
-	}
-}
-
-internal sealed class FieldConverter : JsonConverter<Field>
-{
-	private readonly IElasticsearchClientSettings _settings;
-
-	public FieldConverter(IElasticsearchClientSettings settings) => _settings = settings;
-
-	public override Field? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-	{
-		// TODO - Implement fully
-
-		switch (reader.TokenType)
-		{
-			case JsonTokenType.Null:
-				return null;
-
-			case JsonTokenType.String:
-				return new Field(reader.GetString());
-		}
-
-		reader.Read();
-		return null;
-	}
-
-	public override void Write(Utf8JsonWriter writer, Field value, JsonSerializerOptions options)
-	{
-		if (value is null)
-		{
-			writer.WriteNullValue();
-			return;
-		}
-
-		var fieldName = _settings.Inferrer.Field(value);
-
-		if (string.IsNullOrEmpty(value.Format))
-		{
-			writer.WriteStringValue(fieldName);
-		}
-		else
-		{
-			writer.WriteStartObject();
-			writer.WritePropertyName("field");
-			writer.WriteStringValue(fieldName);
-			writer.WritePropertyName("format");
-			writer.WriteStringValue(value.Format);
-			writer.WriteEndObject();
-		}
-	}
-}
-
 [JsonConverter(typeof(FieldsConverter))]
 [DebuggerDisplay("{DebugDisplay,nq}")]
 public sealed class Fields : IUrlParameter, IEnumerable<Field>, IEquatable<Fields>
