+++ src/Elastic.Clients.Elasticsearch/Common/Infer/JoinFieldRouting/Routing.cs
@@ -174,52 +174,4 @@ namespace Elastic.Clients.Elasticsearch
 		//	return !resolved.IsNullOrEmpty();
 		//}
 	}
-
-	internal sealed class RoutingConverter : JsonConverter<Routing>
-	{
-		private readonly IElasticsearchClientSettings _settings;
-
-		public RoutingConverter(IElasticsearchClientSettings settings) => _settings = settings;
-
-		public override Routing? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) =>
-			reader.TokenType == JsonTokenType.Number
-				? new Routing(reader.GetInt64())
-				: new Routing(reader.GetString());
-
-		public override void Write(Utf8JsonWriter writer, Routing value, JsonSerializerOptions options)
-		{
-			if (value is null)
-			{
-				writer.WriteNullValue();
-				return;
-			}
-
-			if (value.Document is not null)
-			{
-				var documentId = _settings.Inferrer.Routing(value.Document.GetType(), value.Document);
-
-				if (documentId is null)
-				{
-					writer.WriteNullValue();
-					return;
-				}
-
-				writer.WriteStringValue(documentId);
-			}
-			else if (value.DocumentGetter is not null)
-			{
-				var doc = value.DocumentGetter();
-				var documentId = _settings.Inferrer.Routing(doc.GetType(), doc);
-				writer.WriteStringValue(documentId);
-			}
-			else if (value.LongValue.HasValue)
-			{
-				writer.WriteNumberValue(value.LongValue.Value);
-			}
-			else
-			{
-				writer.WriteStringValue(value.StringValue);
-			}
-		}
-	}
 }
